Starting to fix uart comm problem (null bytes)

+++ app/src/sensors/l86_m33/zcbor/l86_m33_encode_types.h
@@ -11,9 +11,9 @@
 #include <stdbool.h>
 #include <stddef.h>
 
-
 #ifdef __cplusplus
-extern "C" {
+extern "C"
+{
 #endif
 
 /** Which value for --default-max-qty this file was created with.
@@ -25,11 +25,13 @@ extern "C" {
  */
 #define DEFAULT_MAX_QTY 3
 
-struct L86_M33 {
-	int64_t timestamp;
-	double latitude;
-	double longitude;
-};
+	struct L86_M33
+	{
+
+		int64_t timestamp;
+		int64_t latitude;
+		int64_t longitude;
+	};
 
 #ifdef __cplusplus
 }
